fix prompt collapse problem'
